cleaning up package dependencies and duplicate station steps
