+++ packages/create-apiops/template/package.json
@@ -6,7 +6,7 @@
     "node": ">=22"
   },
   "dependencies": {
-    "apiops-cycles-method-data": "^3.4.4",
+    "apiops-cycles-method-data": "^4.0.0",
     "canvascreator": "^1.7.3"
   },
   "devDependencies": {
