+++ src/data/method/de/labels.stations.json
@@ -357,8 +357,6 @@
   "station.upskilling.outcomes.1": "Erhöhte Kompetenz bei der Konzeption, Entwicklung und Verwaltung von APIs",
   "station.upskilling.outcomes.2": "Verbesserte Fähigkeit zur Übernahme neuer Technologien und Verfahren",
   "station.upskilling.outcomes.3": "Befähigte Teams zur Förderung von API-Innovation und -Qualität",
-  "station.upskilling.how_it_works.1": "Bereitstellung von Schulungsprogrammen, Workshops und Ressourcen, um Teams bei der Verbesserung ihrer API-Fähigkeiten und -Kenntnisse zu unterstützen.",
-  "station.upskilling.how_it_works.2": "Fördern Sie kontinuierliches Lernen durch Online-Kurse, Zertifizierungen und Engagement in der Gemeinschaft.",
   "station.operating-guidelines.outcomes.1": "Konsistente API-Verwaltung und -Qualität im gesamten Unternehmen",
   "station.operating-guidelines.outcomes.2": "Einhaltung von Organisationsstandards und bewährten Verfahren",
   "station.operating-guidelines.outcomes.3": "Verbesserte Sichtbarkeit und Kontrolle über API-Initiativen",
@@ -367,13 +365,9 @@
   "station.portfolio-management.outcomes.1": "APIs im Einklang mit Unternehmenszielen und strategischen Initiativen",
   "station.portfolio-management.outcomes.2": "Priorisierung und Optimierung von APIs auf der Grundlage von Wert und Wirkung",
   "station.portfolio-management.outcomes.3": "Verbesserte Sichtbarkeit und Kontrolle über das API-Portfolio",
-  "station.portfolio-management.how_it_works.1": "Festlegung von Kriterien für die Bewertung und Priorisierung von APIs auf der Grundlage des Geschäftswerts, der Auswirkungen und der Ausrichtung auf strategische Initiativen.",
-  "station.portfolio-management.how_it_works.2": "Überwachen Sie das API-Portfolio, um sicherzustellen, dass die APIs einen Mehrwert liefern, die Leistungserwartungen erfüllen und mit den Geschäftszielen übereinstimmen.",
   "station.budget-and-resource-management.outcomes.1": "Effiziente Zuweisung von Ressourcen für API-Initiativen",
   "station.budget-and-resource-management.outcomes.2": "Optimierte Investitionen in die API-Entwicklung und -Verwaltung",
   "station.budget-and-resource-management.outcomes.3": "Verbesserte finanzielle Transparenz und Kontrolle über API-Initiativen",
-  "station.budget-and-resource-management.how_it_works.1": "Festlegung von Kriterien für die Bewertung und Priorisierung von Budgetzuweisungen auf der Grundlage des Geschäftswerts, der Auswirkungen und der Ausrichtung auf strategische Initiativen.",
-  "station.budget-and-resource-management.how_it_works.2": "Überwachung des Budgets und der Ressourcennutzung, um effektive Investitionen und Ressourcenzuweisung für API-Initiativen sicherzustellen.",
   "cycle.api-productization-cycle.station.api-product-strategy.title": "API-Produktstrategie - APIs in strategische Produkte verwandeln",
   "cycle.api-productization-cycle.station.api-product-strategy.description": "Bevor Sie etwas bauen, sollten Sie von Anfang an den Wert, die Benutzer und die Geschäftsziele Ihrer API definieren.",
   "cycle.api-productization-cycle.station.api-product-strategy.why_it_matters": "Viele Unternehmen betrachten APIs als technische Projekte und nicht als Produkte. Das Ergebnis? Verwirrte Kunden, schlechte Akzeptanz und verschwendeter Aufwand.\n\nDiese Station hilft Ihnen, den Zweck Ihrer API, die Zielgruppe und die Erfolgskriterien zu definieren, damit Teams APIs bereitstellen können, die echte Probleme lösen.",
