+++ src/data/method/en/labels.stations.json
@@ -339,8 +339,6 @@
   "station.upskilling.outcomes.2": "Improved ability to adopt new technologies and practices",
   "station.upskilling.outcomes.3": "Empowered teams to drive API innovation and quality",
   "station.upskilling.how_it_works": "Identify key API skills and knowledge areas that need enhancement, such as API design, security, and performance.",
-  "station.upskilling.how_it_works.1": "Provide training programs, workshops, and resources to help teams enhance their API skills and knowledge.",
-  "station.upskilling.how_it_works.2": "Encourage continuous learning through online courses, certifications, and community engagement.",
   "station.operating-guidelines.title": "Operating Guidelines - Establish API Governance",
   "station.operating-guidelines.description": "Establish operating guidelines and governance practices to ensure consistent API management and quality.",
   "station.operating-guidelines.why_it_matters": "Effective API governance is essential for ensuring consistent API management, quality, and compliance. This station provides operating guidelines and governance practices that help teams manage APIs effectively, ensuring they align with organizational standards and best practices.",
@@ -361,8 +359,6 @@
   "station.portfolio-management.outcomes.2": "Prioritization and optimization of APIs based on value and impact",
   "station.portfolio-management.outcomes.3": "Improved visibility and control over the API portfolio",
   "station.portfolio-management.how_it_works": "Define portfolio management practices that outline the processes for managing the API portfolio, including prioritization, monitoring, and optimization.",
-  "station.portfolio-management.how_it_works.1": "Establish criteria for evaluating and prioritizing APIs based on business value, impact, and alignment with strategic initiatives.",
-  "station.portfolio-management.how_it_works.2": "Monitor the API portfolio to ensure APIs are delivering value, meeting performance expectations, and aligning with business goals.",
   "station.budget-and-resource-management.title": "Budget and Resource Management - Optimize API Investments",
   "station.budget-and-resource-management.description": "Optimize budget and resource management for API initiatives to ensure effective investments and resource allocation.",
   "station.budget-and-resource-management.why_it_matters": "Effective budget and resource management is essential for ensuring that API initiatives are well-funded and resourced. This station provides guidelines for optimizing budget and resource management, ensuring that API initiatives are aligned with business goals and deliver maximum value.",
@@ -372,8 +368,6 @@
   "station.budget-and-resource-management.outcomes.2": "Optimized investments in API development and management",
   "station.budget-and-resource-management.outcomes.3": "Improved financial visibility and control over API initiatives",
   "station.budget-and-resource-management.how_it_works": "Define budget and resource management practices that outline the processes for managing budgets and resources for API initiatives.",
-  "station.budget-and-resource-management.how_it_works.1": "Establish criteria for evaluating and prioritizing budget allocations based on business value, impact, and alignment with strategic initiatives.",
-  "station.budget-and-resource-management.how_it_works.2": "Monitor budget and resource utilization to ensure effective investments and resource allocation for API initiatives.",
   "cycle.api-productization-cycle.station.api-product-strategy.title": "API Product Strategy - Turn APIs Into Strategic Products",
   "cycle.api-productization-cycle.station.api-product-strategy.description": "Before building anything, define your API's value, users, and business goals from day one.",
   "cycle.api-productization-cycle.station.api-product-strategy.why_it_matters": "Many organizations think of APIs as tech projects, not products. The result? Confused consumers, poor adoption, and wasted effort.\n\nThis station helps you define your APIâ€™s purpose, target audience, and success criteria, so teams can deliver APIs that solve real problems.",
