+++ src/data/method/fi/labels.stations.json
@@ -357,8 +357,6 @@
   "station.upskilling.outcomes.1": "API-suunnittelun, -kehityksen ja -hallinnan lisääntynyt osaaminen.",
   "station.upskilling.outcomes.2": "Parempi kyky omaksua uutta teknologiaa ja uusia käytäntöjä",
   "station.upskilling.outcomes.3": "Valtuutetaan tiimit edistämään API-innovaatioita ja laatua.",
-  "station.upskilling.how_it_works.1": "Tarjotaan koulutusohjelmia, työpajoja ja resursseja, joiden avulla tiimit voivat parantaa API-taitojaan ja -tietämystään.",
-  "station.upskilling.how_it_works.2": "Kannustetaan jatkuvaan oppimiseen verkkokurssien, sertifikaattien ja yhteisöllisen sitoutumisen avulla.",
   "station.operating-guidelines.outcomes.1": "Yhdenmukainen API-hallinta ja laatu koko organisaatiossa",
   "station.operating-guidelines.outcomes.2": "organisaation standardien ja parhaiden käytäntöjen noudattaminen",
   "station.operating-guidelines.outcomes.3": "Parempi näkyvyys ja API-aloitteiden hallinta",
@@ -367,13 +365,9 @@
   "station.portfolio-management.outcomes.1": "APIt linjassa liiketoiminnan tavoitteiden ja strategisten aloitteiden kanssa.",
   "station.portfolio-management.outcomes.2": "APIen priorisointi ja optimointi arvon ja vaikutusten perusteella.",
   "station.portfolio-management.outcomes.3": "Parempi näkyvyys ja API-portfolion hallinta",
-  "station.portfolio-management.how_it_works.1": "Laaditaan kriteerit rajapintojen arvioimiseksi ja priorisoimiseksi liiketoiminta-arvon, vaikutusten ja strategisten aloitteiden kanssa yhteensopivuuden perusteella.",
-  "station.portfolio-management.how_it_works.2": "Seuraa API-portfoliota varmistaaksesi, että APIt tuottavat arvoa, täyttävät suorituskykyodotukset ja ovat linjassa liiketoiminnan tavoitteiden kanssa.",
   "station.budget-and-resource-management.outcomes.1": "Resurssien tehokas kohdentaminen API-aloitteisiin",
   "station.budget-and-resource-management.outcomes.2": "Optimoidut investoinnit API-kehitykseen ja -hallintaan",
   "station.budget-and-resource-management.outcomes.3": "Parempi taloudellinen näkyvyys ja API-aloitteiden valvonta.",
-  "station.budget-and-resource-management.how_it_works.1": "Laaditaan kriteerit budjettimäärärahojen arvioimiseksi ja priorisoimiseksi liiketoiminta-arvon, vaikutusten ja strategisten aloitteiden mukaisuuden perusteella.",
-  "station.budget-and-resource-management.how_it_works.2": "Seuraa budjettia ja resurssien käyttöä varmistaaksesi tehokkaat investoinnit ja resurssien jakamisen API-aloitteisiin.",
   "cycle.api-productization-cycle.station.api-product-strategy.title": "API-tuotestrategia - Tee rajapinnoista strategisia tuotteita",
   "cycle.api-productization-cycle.station.api-product-strategy.description": "Ennen kuin rakennat mitään, määrittele APIn arvo, hyödyntäjät ja liiketoiminnalliset tavoitteet heti alusta alkaen.",
   "cycle.api-productization-cycle.station.api-product-strategy.why_it_matters": "Monet organisaatiot pitävät rajapintoja teknisinä projekteina, eivät tuotteina. Tulos? Hämmentyneitä hyödyntäjiä, heikkoa käyttöönottoa ja turhaa työtä. Tämä asema auttaa määrittelemään APIn tarkoituksen, kohderyhmän ja menestyskriteerit, jotta tiimit voivat tuottaa rajapintoja, jotka ratkaisevat todellisia ongelmia.",
