+++ src/data/method/fr/labels.stations.json
@@ -357,8 +357,6 @@
   "station.upskilling.outcomes.1": "Compétences accrues en matière de conception, de développement et de gestion des API",
   "station.upskilling.outcomes.2": "Amélioration de la capacité à adopter de nouvelles technologies et pratiques",
   "station.upskilling.outcomes.3": "Des équipes responsabilisées pour favoriser l'innovation et la qualité de l'API",
-  "station.upskilling.how_it_works.1": "Fournir des programmes de formation, des ateliers et des ressources pour aider les équipes à améliorer leurs compétences et leurs connaissances en matière d'API.",
-  "station.upskilling.how_it_works.2": "Encourager la formation continue par des cours en ligne, des certifications et l'engagement communautaire.",
   "station.operating-guidelines.outcomes.1": "Gestion et qualité cohérentes de l'API dans l'ensemble de l'organisation",
   "station.operating-guidelines.outcomes.2": "Respect des normes et des meilleures pratiques de l'organisation",
   "station.operating-guidelines.outcomes.3": "Amélioration de la visibilité et du contrôle des initiatives en matière d'API",
@@ -367,13 +365,9 @@
   "station.portfolio-management.outcomes.1": "Des API alignées sur les objectifs de l'entreprise et les initiatives stratégiques",
   "station.portfolio-management.outcomes.2": "Hiérarchisation et optimisation des API en fonction de leur valeur et de leur impact",
   "station.portfolio-management.outcomes.3": "Amélioration de la visibilité et du contrôle du portefeuille d'API",
-  "station.portfolio-management.how_it_works.1": "Établir des critères d'évaluation et de hiérarchisation des API en fonction de la valeur métier, de l'impact et de l'alignement sur les initiatives stratégiques.",
-  "station.portfolio-management.how_it_works.2": "Contrôler le portefeuille d'API pour s'assurer que les API apportent de la valeur, répondent aux attentes en matière de performances et s'alignent sur les objectifs de l'entreprise.",
   "station.budget-and-resource-management.outcomes.1": "Affectation efficace des ressources aux initiatives de l'API",
   "station.budget-and-resource-management.outcomes.2": "Optimisation des investissements dans le développement et la gestion des API",
   "station.budget-and-resource-management.outcomes.3": "Amélioration de la visibilité financière et du contrôle des initiatives API",
-  "station.budget-and-resource-management.how_it_works.1": "Établir des critères d'évaluation et de hiérarchisation des affectations budgétaires en fonction de la valeur métier, de l'impact et de l'alignement sur les initiatives stratégiques.",
-  "station.budget-and-resource-management.how_it_works.2": "Contrôler le budget et l'utilisation des ressources afin de garantir l'efficacité des investissements et de l'affectation des ressources pour les initiatives de l'API.",
   "cycle.api-productization-cycle.station.api-product-strategy.title": "Stratégie produit API - Transformer les API en produits stratégiques",
   "cycle.api-productization-cycle.station.api-product-strategy.description": "Avant de construire quoi que ce soit, définissez dès le départ la valeur de votre API, ses utilisateurs et ses objectifs métier.",
   "cycle.api-productization-cycle.station.api-product-strategy.why_it_matters": "De nombreuses organisations considèrent les API comme des projets technologiques et non comme des produits. Le résultat ? Des utilisateurs désorientés, une adoption médiocre et un gaspillage d'efforts.\n\nCette station vous aide à définir l'objectif, le public cible et les critères de réussite de votre API, afin que les équipes puissent fournir des API qui résolvent des problèmes réels.",
