+++ src/data/method/pt/labels.stations.json
@@ -339,8 +339,6 @@
   "station.upskilling.outcomes.2": "Melhoria da capacidade de adoção de novas tecnologias e práticas",
   "station.upskilling.outcomes.3": "Equipas capacitadas para impulsionar a inovação e a qualidade da API",
   "station.upskilling.how_it_works": "Identificar as principais competências e áreas de conhecimento da API que precisam de ser melhoradas, como a conceção, a segurança e o desempenho da API.",
-  "station.upskilling.how_it_works.1": "Fornecer programas de formação, workshops e recursos para ajudar as equipas a melhorar as suas competências e conhecimentos em matéria de API.",
-  "station.upskilling.how_it_works.2": "Incentivar a aprendizagem contínua através de cursos em linha, certificações e envolvimento da comunidade.",
   "station.operating-guidelines.title": "Diretrizes operacionais - Estabelecer a governação da API",
   "station.operating-guidelines.description": "Estabelecer diretrizes operacionais e práticas de governação para garantir uma gestão e qualidade consistentes da API.",
   "station.operating-guidelines.why_it_matters": "A governação eficaz da API é essencial para garantir uma gestão, qualidade e conformidade consistentes da API. Esta estação fornece diretrizes operacionais e práticas de governação que ajudam as equipas a gerir as APIs de forma eficaz, assegurando o seu alinhamento com as normas organizacionais e as boas práticas.",
@@ -361,8 +359,6 @@
   "station.portfolio-management.outcomes.2": "Priorização e otimização de APIs com base no valor e no impacto",
   "station.portfolio-management.outcomes.3": "Melhoria da visibilidade e do controlo da carteira de API",
   "station.portfolio-management.how_it_works": "Definir práticas de gestão da carteira que descrevam os processos de gestão da carteira de API, incluindo a definição de prioridades, a monitorização e a otimização.",
-  "station.portfolio-management.how_it_works.1": "Estabelecer critérios para avaliar e dar prioridade às APIs com base no valor de negócio, no impacto e no alinhamento com iniciativas estratégicas.",
-  "station.portfolio-management.how_it_works.2": "Monitorizar a carteira de APIs para garantir que as APIs estão a fornecer valor, a cumprir as expectativas de desempenho e a alinhar-se com os objetivos de negócio.",
   "station.budget-and-resource-management.title": "Orçamento e gestão de recursos - Otimizar os investimentos em API",
   "station.budget-and-resource-management.description": "Otimizar o orçamento e a gestão de recursos para as iniciativas API, a fim de garantir investimentos e atribuição de recursos eficazes.",
   "station.budget-and-resource-management.why_it_matters": "A gestão eficaz do orçamento e dos recursos é essencial para garantir que as iniciativas de API sejam bem financiadas e dotadas de recursos. Esta estação fornece diretrizes para otimizar a gestão do orçamento e dos recursos, assegurando que as iniciativas de API estão alinhadas com os objetivos de negócio e proporcionam o máximo valor.",
@@ -372,8 +368,6 @@
   "station.budget-and-resource-management.outcomes.2": "Investimentos optimizados no desenvolvimento e gestão de API",
   "station.budget-and-resource-management.outcomes.3": "Melhoria da visibilidade financeira e do controlo das iniciativas API",
   "station.budget-and-resource-management.how_it_works": "Definir práticas de gestão orçamental e de recursos que descrevam os processos de gestão de orçamentos e recursos para iniciativas de API.",
-  "station.budget-and-resource-management.how_it_works.1": "Estabelecer critérios para avaliar e dar prioridade às dotações orçamentais com base no valor de negócio, no impacto e no alinhamento com as iniciativas estratégicas.",
-  "station.budget-and-resource-management.how_it_works.2": "Controlar o orçamento e a utilização dos recursos para garantir investimentos eficazes e a afetação de recursos às iniciativas da API.",
   "cycle.api-productization-cycle.station.api-product-strategy.title": "Estratégia de produtos API - Transforme APIs em produtos estratégicos",
   "cycle.api-productization-cycle.station.api-product-strategy.description": "Antes de construir qualquer coisa, defina o valor da sua API, os utilizadores e os objetivos de negócio desde o primeiro dia.",
   "cycle.api-productization-cycle.station.api-product-strategy.why_it_matters": "Muitas organizações pensam nas APIs como projetos tecnológicos e não como produtos. O resultado? Consumidores confusos, fraca adoção e esforço desperdiçado.\n\nEsta estação ajuda-o a definir o objetivo da sua API, o público-alvo e os critérios de sucesso, para que as equipas possam fornecer APIs que resolvam problemas reais.",
